# Conflicts:
#	src/main/java/org/blank/projectmanagementsystem/ProjectManagementSystemApplication.java
#	src/main/java/org/blank/projectmanagementsystem/controller/MainController.java
#	src/main/java/org/blank/projectmanagementsystem/controller/MemberController.java
#	src/main/java/org/blank/projectmanagementsystem/domain/entity/Issue.java
#	src/main/java/org/blank/projectmanagementsystem/domain/entity/Project.java
#	src/main/java/org/blank/projectmanagementsystem/domain/entity/ResponsibleParty.java
#	src/main/java/org/blank/projectmanagementsystem/domain/entity/User.java
#	src/main/java/org/blank/projectmanagementsystem/repository/UserRepository.java
#	src/main/java/org/blank/projectmanagementsystem/service/ProjectService.java
#	src/main/java/org/blank/projectmanagementsystem/service/impl/ProjectServiceImpl.java
#	src/main/resources/application.properties
#	src/main/resources/static/css/hope.css
#	src/main/resources/static/css/style.css
#	src/test/java/org/blank/projectmanagementsystem/domain/entity/IssueTest.java

+++ src/main/java/org/blank/projectmanagementsystem/repository/IssuePlaceRepository.java
@@ -3,5 +3,5 @@ package org.blank.projectmanagementsystem.repository;
 import org.blank.projectmanagementsystem.domain.entity.IssuePlace;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IssuePlaceRepository extends JpaRepository<IssuePlace, Integer> {
+public interface IssuePlaceRepository extends JpaRepository<IssuePlace, Long> {
 }
